Added Scroll Event to Show Current Section of Page on Navigation Bar
Modified 'index.html', style.css, and wrote into 'main.js' to have a marker show up under each navigation link in the navbar when it's respective section is in the viewport. This feature is still unfinished and buggy. More code is required in 'main.js'. Lines of code utilised for testing have been left commented, incase of their future need.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>