Completed Script for Scroll Event to Indicate Current Section of Page on Navigation Bar
Modified the script in 'main.js' that creates a marker under each navigation link in the navbar when it's respective section is in the viewport. This feature is now complete and fully functional. Lines of code utilised for testing have been still been left commented, incase of their future need.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>