Wrote JavaScript Script in 'main.js' to Implement 'flicker' Animation
Wrote code in 'main.js' that places the 'flicker' class which holds the styling for the 'flicker' animation into the header of whichever section is in the viewport. This works when using the navigation buttons to go to sections and when scrolling through sections. Also added and modified styling of teh section headers to ensure their overall look, pre-animation, matches the animation. Furthermore, there was a bug in one of the media queries found in 'style.css' which was stopping the grids from changing to a single column on mobile. This was fixed.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>