Improved Effect of Flickering Animation Trigger Script in 'main.js'
Modified the script in 'main.js' which provides functionality for triggering the 'flicker' animation on the header of the section that comes into view. Now the 'flicker' animation triggers closer to the top of the section, which means there is less of chance to accidentally it while not having the section header in view. This ensures a better user experience.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>