Massive UI Theming Changes and Contact Form Completed
Big update. The UI of the entire website was modified to integrate a neon lights theme. This was done because the neon light 'flicker' animation headers of each section looked out of place and they were important to the overall look of the site that I wanted. Thus, I had to integrate them into the whole UI seemlessly, by modifying various UI elements to follow a neon lights theme. Furthermore, the contact form was completed, in the 'contact' section. This was tested with various styling and is the final product.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>