Fixed Scrolling Script Bugs in 'main.js' for Mobile Devices
Due to the fact that each section is a different height and greater than 100 viewport height on mobile, it is causing in consistencies with how the scrolling algorithm works. I have implemented a quick temporary fix to the problem for now but this requires the website to be completed before a permanent fix is available.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>