Redesigned the 'contact' Section
The profile shot was removed from the contact section and will be added to another section (most likely 'skills' section will be changed to 'about' and image will be added there). Furthermore, buttons were added for direct email and calling. The grid for the 'contact' section was also modified slightly.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>