Scroll Event Bug Fix in 'main.js'
There was a big with the 'onscroll' event listener script in 'main.js'. The problem was that the script was using the the moment when the top of the viewport intersected the [top of each section - 1/3 of the section height] as a key-point to perform specific actions for the site, such as indicating the active section in the navbar. The problem with this was that if a section was bigger than 2 times the viewport height (which was the case for mobile users), the intersection point of that specific section came before the section came into the viewport. This meant that the navbar on mobile was very buggy. I fixed this by simply making the point of intersection between the top of the viewport and the [top of each section - 2/5 of the viewport height] as the key-point to perform all the necessary actions that the 'onscroll' event script has to do.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>