Addition of a Skills Subsection in the 'About' Section
I designed and added in a skills subsection into my 'About' section as I felt that I needed to have my skillset in my website. Originally, the 'About' section was the 'Skills' section but I felt that my skills section was too small and bland. Now I realize that the best thing to do is to have my skills listed in my 'About' section. To make my skills seem less boring to look at I designed a grid design which will showcase my skills in a modular/categorized layout. So far, only the grid template itself has been completed and is fully responsive. The content for the grid is to come next, inshallah.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>