Created Graphics for Past Projects and Added them to 'Work' Section
Created graphics for past projects on Photoshop and then added them to the respective places in the 'Work' section. I also utilized the media queries to scale the height, length, and absolute/relative placement of these images according to the Bootstrap defined respective screen size breakpoints. The 'Work' section was created with future upgradability in mind and will only require you to add new projects in the same format as the current ones in the 'index.html' file and the required styling will be applied to them as well.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>