The UI of the Website has been Completed.
The website's UI elements are done. There are some features that I have thought of implementing for a future version, such as a parallax scrolling background and a carousel listing style for each tile in the skills subsection grid. There are still some broken/unreferenced buttons and links on the website and the contact form is not yet functional.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>