Added Mockup for C Programs Card in 'Work' Section
Created a mockup of some of the code from the C programs repository found on github, using photoshop. Uploaded the mockup to the website and added it in to it's respective place using css in 'style.css'. Also added title for the C programs card, found in the 'Work' section, in 'index.html'.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>